gmail_agent/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# LangChain & LangGraph
langgraph==0.2.50
langchain==0.3.12
langchain-core==0.3.21
langchain-openai==0.2.12

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Optional: For better performance
orjson==3.9.15
